Added controller level authorization to OrderService

+++ OrderService/src/main/kotlin/it/polito/wa2/orderservice/controllers/OrderController.kt
@@ -1,11 +1,13 @@
 package it.polito.wa2.orderservice.controllers
 
+import it.polito.wa2.orderservice.constants.RoleName
 import it.polito.wa2.orderservice.constants.Values
 import it.polito.wa2.orderservice.dtos.OrderDTO
 import it.polito.wa2.orderservice.services.interfaces.OrderService
 import it.polito.wa2.orderservice.validators.validatePatch
 import it.polito.wa2.orderservice.validators.validatePost
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.data.domain.Page
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -22,9 +24,19 @@ class OrderController {
     fun getOrders(
         @RequestParam("pageNo", defaultValue = Values.DEFAULT_PAGE_NO) @Min(0) pageNo: Int,
         @RequestParam("pageSize", defaultValue = Values.DEFAULT_PAGE_SIZE) @Min(1) pageSize: Int,
+        @RequestHeader("userId") userId: String?,
+        @RequestHeader("roles") roles: String?
     ) : ResponseEntity<Any> {
         try {
-            val orderPageDTO = orderService.getOrders(pageNo, pageSize)
+            val orderPageDTO: Page<OrderDTO>
+            if (roles == null)
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
+            if (roles.contains(RoleName.ROLE_ADMIN.value))
+                orderPageDTO = orderService.getOrders(pageNo, pageSize)
+            else if (roles.contains(RoleName.ROLE_CUSTOMER.value) && userId != null)
+                orderPageDTO = orderService.getCustomerOrders(userId, pageNo, pageSize)
+            else
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
             val response = hashMapOf<String, Any>()
             response["orders"] = orderPageDTO.content
             response["currentPage"] = pageNo
@@ -38,11 +50,25 @@ class OrderController {
 
     @GetMapping("/{orderId}")
     fun getOrderById(
-        @PathVariable("orderId") orderId: Long
+        @PathVariable("orderId") orderId: Long,
+        @RequestHeader("userId") userId: String?,
+        @RequestHeader("roles") roles: String?
     ): ResponseEntity<Any> {
         try {
-            val orderDTO = orderService.getOrderById(orderId)
-            return ResponseEntity.ok(orderDTO)
+            if (roles == null)
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
+            if (roles.contains(RoleName.ROLE_ADMIN.value)) {
+                val orderDTO = orderService.getOrderById(orderId)
+                return ResponseEntity.ok(orderDTO)
+            }
+            else if (roles.contains(RoleName.ROLE_CUSTOMER.value) && userId != null) {
+                val orderDTO = orderService.getOrderById(orderId)
+                if (orderDTO.userId != userId.toLong())
+                    return ResponseEntity.status(401).body(Values.UNAUTHORIZED)
+                return ResponseEntity.ok(orderDTO)
+            }
+            else
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
         } catch(e: RuntimeException) {
             return ResponseEntity.badRequest().body(e.message)
         }
@@ -51,11 +77,26 @@ class OrderController {
     @PostMapping
     fun createOrder(
         @RequestBody orderDTO: OrderDTO,
+        @RequestHeader("userId") userId: String?,
+        @RequestHeader("roles") roles: String?
     ): ResponseEntity<Any> {
         try {
-            if (!orderDTO.validatePost()) return ResponseEntity.badRequest().body(Values.INVALID_ORDER_REPRESENTATION)
-            val responseOrderDTO = orderService.createOrder(orderDTO)
-            return ResponseEntity.status(HttpStatus.CREATED).body(responseOrderDTO)
+            if (roles == null)
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
+            if (!orderDTO.validatePost())
+                return ResponseEntity.badRequest().body(Values.INVALID_ORDER_REPRESENTATION)
+            if (roles.contains(RoleName.ROLE_ADMIN.value)) {
+                val responseOrderDTO = orderService.createOrder(orderDTO)
+                return ResponseEntity.status(HttpStatus.CREATED).body(responseOrderDTO)
+            }
+            else if (roles.contains(RoleName.ROLE_CUSTOMER.value) && userId != null) {
+                if (orderDTO.userId != userId.toLong())
+                    return ResponseEntity.status(401).body(Values.UNAUTHORIZED)
+                val responseOrderDTO = orderService.createOrder(orderDTO)
+                return ResponseEntity.ok(responseOrderDTO)
+            }
+            else
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
         } catch (e: RuntimeException) {
             return ResponseEntity.badRequest().body(e.message)
         }
@@ -64,9 +105,15 @@ class OrderController {
     @PatchMapping("/{orderId}")
     fun updateOrder(
         @PathVariable("orderId") orderId: Long,
-        @RequestBody orderDTO: OrderDTO
+        @RequestBody orderDTO: OrderDTO,
+        @RequestHeader("userId") userId: String?,
+        @RequestHeader("roles") roles: String?
     ): ResponseEntity<Any> {
         try {
+            if (roles == null)
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
+            if (!roles.contains(RoleName.ROLE_ADMIN.value))
+                return ResponseEntity.status(401).body(Values.UNAUTHORIZED)
             if (!orderDTO.validatePatch()) return ResponseEntity.badRequest().body(Values.INVALID_ORDER_REPRESENTATION)
             val responseOrderDTO = orderService.updateOrder(orderId, orderDTO)
             return ResponseEntity.ok(responseOrderDTO)
@@ -76,12 +123,27 @@ class OrderController {
     }
 
     @DeleteMapping("/{orderId}")
-    fun deleteProduct(
-        @PathVariable("orderId") orderId: Long
+    fun deleteOrder(
+        @PathVariable("orderId") orderId: Long,
+        @RequestHeader("userId") userId: String?,
+        @RequestHeader("roles") roles: String?
     ): ResponseEntity<Any> {
         try {
-            orderService.deleteOrder(orderId)
-            return ResponseEntity.noContent().build()
+            if (roles == null)
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
+            if (roles.contains(RoleName.ROLE_ADMIN.value)) {
+                orderService.deleteOrder(orderId)
+                return ResponseEntity.noContent().build()
+            }
+            else if (roles.contains(RoleName.ROLE_CUSTOMER.value) && userId != null) {
+                val orderDTO = orderService.getOrderById(orderId)
+                if (orderDTO.userId != userId.toLong())
+                    return ResponseEntity.status(401).body(Values.UNAUTHORIZED)
+                orderService.deleteOrder(orderId)
+                return ResponseEntity.noContent().build()
+            }
+            else
+                return ResponseEntity.badRequest().body(Values.FAILED_TO_AUTHORIZE)
         } catch (e: RuntimeException) {
             return ResponseEntity.badRequest().body(e.message)
         }
